src/backend/gateway/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
fastapi==0.118.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
passlib==1.7.4
pillow==11.3.0
pycryptodome==3.23.0
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
python-multipart==0.0.20
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0